"/adicionarItem" Concluido

+++ API - De Little a Big Açaí/de-litte-a-big-acai/src/main/java/com/litte_acai/de_litte_a_big_acai/service/ItemService.java
@@ -14,7 +14,7 @@ public interface ItemService {
                         String descricaoItem,
                         String categoria,
                         Double precoUni,
-                        Integer quant,
+                        Double quant,
                         Double volumeUni,
                         String unidMedida,
                         LocalDate dataValidadeLocalDate,
